finance-agent-team/requirements.txt
phidata
ollama
duckduckgo-search
yfinance
sqlalchemy
fastapi[standard]
python-docx